OK my bad - Options Menu
The code I get from the GitHub app isn't the one that gets updated whenever I make changes in unity. This is because when I run code in unity it opens Visual Studio(Purple) and not visual Studio code(Blue). So whenever I finish some work I have to manually transfer the files over to this branch and well this time I ended up deleting everything.

So I tried to copy over the other files from unity, all the files.